pyproject.toml
[tool.poetry]
name = "pyfront"
version = "0.1.0"
description = ""
authors = ["Nikita Irgashev <nik.irg@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
websockets = "^12.0"
uvicorn = "^0.24.0.post1"
pydantic = "^2.6.4"
beanie = "^1.25.0"
python-multipart = "^0.0.9"
starlette-sessions = "^0.2.0"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
